EN-2184 Changed Swagger implementation to use Password Grant type

+++ engine/src/main/java/org/entando/entando/web/user/UserController.java
@@ -36,6 +36,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.DataBinder;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -48,6 +49,7 @@ import java.util.Map;
  *
  * @author paddeo
  */
+@Validated
 @RestController
 @RequestMapping(value = "/users")
 @SessionAttributes("user")
@@ -150,6 +152,11 @@ public class UserController {
     @RequestMapping(value = "/{username:.+}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> deleteUser(@PathVariable String username) throws ApsSystemException {
         logger.debug("deleting {}", username);
+
+        if (!userValidator.isValidDeleteUser(username)) {
+            throw new ApsSystemException("Sorry. You can't delete the administrator user");
+        }
+
         this.getUserService().removeUser(username);
         Map<String, String> result = new HashMap<>();
         result.put("code", username);
